server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pynacl==1.5.0
httpx==0.28.0
brotli==1.1.0
dnspython==2.7.0
python-whois==0.9.4
psycopg[binary]==3.2.0
redis==5.2.0
pydantic==2.10.0
pydantic-settings==2.6.0
